requirements.txt
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic
chromadb
fastapi
uvicorn
python-dotenv
requests
groq